requirements.txt
# Основные зависимости проекта
# Добавляйте новые зависимости по мере необходимости 
PyQt6==6.6.1        # Графический интерфейс
Pillow==10.2.0      # Работа с изображениями
cairosvg==2.7.1     # Поддержка SVG 